fix(stream): fix syntax error.

+++ source/dnode/vnode/src/tqCommon/tqCommon.c
@@ -905,7 +905,7 @@ int32_t tqStreamTaskProcessRetrieveTriggerReq(SStreamMeta* pMeta, SRpcMsg* pMsg)
     streamTaskGetActiveCheckpointInfo(pTask, &transId, &checkpointId);
     if (checkpointId != pReq->checkpointId) {
       tqError("s-task:%s invalid checkpoint-trigger retrieve msg from %x, current checkpointId:%"PRId64" req:%"PRId64,
-          pTask->id.idStr, pReq->downstreamTaskId, checkpointId, pReq->checkpointId);
+          pTask->id.idStr, (int32_t) pReq->downstreamTaskId, checkpointId, pReq->checkpointId);
       streamMetaReleaseTask(pMeta, pTask);
       return TSDB_CODE_INVALID_MSG;
     }
